iio: jesd204: Add library for reconfiguring Xilinx high-speed transceiver
The Xilinx high-speed transceivers feature a so called
dynamic-reconfiguration-port (DRP) that can be used to reconfigure the
operating parameters of the transceiver at runtime.

Add a standalone library module that can be used to calculate and program
the DRP settings that are relevant for JESD204 operation. It can be used by
drivers that need to re-program the high-speed transceivers to handle the
low level details.

The module is independent of the access scheme that is used for the DRP
registers and provides a set of drp_ops with read and write callbacks that
need to filled in by a driver making use of the library.

Based on the existing axi_adxcvr and axi_jesd204b_gt drivers.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>